Insert new movie, integration tests, refactor, checkstyle



See merge request !8

+++ src/main/java/application/integration/dbsakila/entity/LanguageEntity.java
@@ -15,17 +15,17 @@ import javax.persistence.Table;
 public class LanguageEntity {
 
     @Id
-    @GeneratedValue(strategy= GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "language_id")
-    private int id;
+    private long id;
 
     private String name;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
